StashRepository: Only check pull request when canceling or aborting build
The caller is supposed to ensure that the build being canceled or aborted
corresponds to the job that is scheduling a new build. That is a stronger
condition than the source repository match. The job and the pull request
ID define the pull request uniquely, and the pull request is associated
with one source repository.